Mutable per target channel filters

+++ telemirror/mirroring.py
@@ -1,12 +1,11 @@
 import logging
 from typing import Dict, List, Union
 
+from config import TargetConfig
 from telethon import TelegramClient, events, utils
 from telethon.extensions import markdown
 from telethon.sessions import StringSession
 from telethon.tl import types
-
-from config import Config
 
 from .hints import EventLike, MessageLike
 from .storage import Database, MirrorMessage
@@ -38,15 +37,9 @@ class EventHandlers:
         incoming_message: MessageLike = event.message
 
         try:
-            config = self._mirror_config.get(incoming_chat_id)
-
-            outgoing_chats = config.to
+            outgoing_chats = self._chat_mapping.get(incoming_chat_id)
             if not outgoing_chats:
                 self._logger.warning(f'No target chats for {incoming_chat_id}')
-                return
-
-            if await config.filters.process(incoming_message) is False:
-                self._logger.info(f'Skipping message {incoming_message_link} by filter')
                 return
 
             reply_to_messages: dict[int, int] = {
@@ -54,21 +47,30 @@ class EventHandlers:
                 for m in await self._database.get_messages(incoming_message.reply_to_msg_id, incoming_chat_id)
             } if incoming_message.is_reply else {}
 
+            if isinstance(incoming_message.media, types.MessageMediaPoll):
+                # Copy quiz poll as simple poll
+                incoming_message.media.poll.quiz = None
+
             for outgoing_chat in outgoing_chats:
 
-                if isinstance(incoming_message.media, types.MessageMediaPoll):
-                    # Copy quiz poll as simple poll
-                    incoming_message.media.poll.quiz = None
+                config = self._target_config.get(outgoing_chat)
+
+                proceed, filtered_message = await config.filters.process(incoming_message)
+
+                if proceed is False:
+                    self._logger.info(
+                        f'Skipping message {incoming_message_link} for {outgoing_chat}')
+                    continue
 
                 outgoing_message = await self.send_message(
                     entity=outgoing_chat,
-                    message=incoming_message,
-                    formatting_entities=incoming_message.entities,
+                    message=filtered_message,
+                    formatting_entities=filtered_message.entities,
                     reply_to=reply_to_messages.get(outgoing_chat)
                 )
 
                 if outgoing_message:
-                    await self._database.insert(MirrorMessage(original_id=incoming_message.id,
+                    await self._database.insert(MirrorMessage(original_id=filtered_message.id,
                                                               original_channel=incoming_chat_id,
                                                               mirror_id=outgoing_message.id,
                                                               mirror_channel=outgoing_chat))
@@ -95,26 +97,10 @@ class EventHandlers:
         incoming_chat_id: int = event.chat_id
 
         try:
-            config = self._mirror_config.get(incoming_chat_id)
-
-            outgoing_chats = config.to
+            outgoing_chats = self._chat_mapping.get(incoming_chat_id)
             if not outgoing_chats:
                 self._logger.warning(f'No target chats for {incoming_chat_id}')
                 return
-
-            # Apply filters to first non-empty or first message
-            if await config.filters.process(next((m for m in incoming_album if m.message), incoming_first_message)) is False:
-                self._logger.info(f'Skipping album {incoming_message_link} by filter')
-                return
-
-            idx: list[int] = []
-            files: list[types.TypeMessageMedia] = []
-            captions: list[str] = []
-            for incoming_message in incoming_album:
-                idx.append(incoming_message.id)
-                files.append(incoming_message.media)
-                # Pass unparsed text, since: https://github.com/LonamiWebs/Telethon/issues/3065
-                captions.append(incoming_message.text)
 
             reply_to_messages: dict[int, int] = {
                 m.mirror_channel: m.mirror_id
@@ -122,6 +108,40 @@ class EventHandlers:
             } if incoming_first_message.is_reply else {}
 
             for outgoing_chat in outgoing_chats:
+                config = self._target_config.get(outgoing_chat)
+
+                proceed = True
+                filtered = False
+                filtered_album: list[MessageLike] = []
+
+                for m in incoming_album:
+                    if m.message and not filtered:
+                        filtered = True
+                        proceed, filtered_message = await config.filters.process(m)
+                        if proceed is False:
+                            break
+                        filtered_album.append(filtered_message)
+                    else:
+                        filtered_album.append(m)
+
+                if not filtered:
+                    proceed, filtered_album[0] = await config.filters.process(filtered_album[0])
+
+                # Apply filters to first non-empty or first message
+                if proceed is False:
+                    self._logger.info(
+                        f'Skipping album {incoming_message_link} for {outgoing_chat}')
+                    continue
+
+                idx: list[int] = []
+                files: list[types.TypeMessageMedia] = []
+                captions: list[str] = []
+                for incoming_message in filtered_album:
+                    idx.append(incoming_message.id)
+                    files.append(incoming_message.media)
+                    # Pass unparsed text, since: https://github.com/LonamiWebs/Telethon/issues/3065
+                    captions.append(incoming_message.text)
+
                 outgoing_messages: list[types.Message] = await self.send_file(
                     entity=outgoing_chat,
                     caption=captions,
@@ -149,13 +169,8 @@ class EventHandlers:
 
         incoming_message: MessageLike = event.message
         incoming_chat: int = event.chat_id
-
-        config = self._mirror_config.get(incoming_chat)
-
-        if config.disable_edit is True:
-            return
-
         incoming_message_link: str = self.event_message_link(event)
+
         self._logger.info(f'Edit message: {incoming_message_link}')
 
         try:
@@ -165,18 +180,26 @@ class EventHandlers:
                     f'No target messages to edit for {incoming_message_link}')
                 return
 
-            if incoming_message.grouped_id is None or incoming_message.message:
-                await config.filters.process(incoming_message)
-
             for outgoing_message in outgoing_messages:
+                config = self._target_config.get(
+                    outgoing_message.mirror_channel)
+
+                if config.disable_edit is True:
+                    continue
+
+                filtered_message = incoming_message
+
+                if incoming_message.grouped_id is None or incoming_message.message:
+                    _, filtered_message = await config.filters.process(incoming_message)
+
                 await self.edit_message(
                     entity=outgoing_message.mirror_channel,
                     message=outgoing_message.mirror_id,
-                    text=incoming_message.message,
-                    formatting_entities=incoming_message.entities,
-                    file=incoming_message.media,
+                    text=filtered_message.message,
+                    formatting_entities=filtered_message.entities,
+                    file=filtered_message.media,
                     link_preview=isinstance(
-                        incoming_message.media, types.MessageMediaWebPage)
+                        filtered_message.media, types.MessageMediaWebPage)
                 )
         except Exception as e:
             self._logger.error(e, exc_info=True)
@@ -186,9 +209,6 @@ class EventHandlers:
 
         deleted_ids: List[int] = event.deleted_ids
         incoming_chat: int = event.chat_id
-
-        if self._mirror_config.get(incoming_chat).disable_delete is True:
-            return
 
         self._logger.info(
             f'Delete {len(deleted_ids)} messages from {incoming_chat}')
@@ -203,6 +223,12 @@ class EventHandlers:
 
                 for deleting_message in deleting_messages:
                     try:
+                        config = self._target_config.get(
+                            deleting_message.mirror_channel)
+
+                        if config.disable_delete is True:
+                            continue
+
                         await self.delete_messages(
                             entity=deleting_message.mirror_channel,
                             message_ids=deleting_message.mirror_id
@@ -232,19 +258,22 @@ class Mirroring(EventHandlers):
 
     def __init__(
         self: 'MirrorTelegramClient',
-        mirror_config: Dict[int, Config],
+        target_config: Dict[int, TargetConfig],
+        chat_mapping: Dict[int, List[int]],
         database: Database,
         logger: Union[str, logging.Logger] = None,
     ) -> None:
         """Configure channels mirroring
 
         Args:
-            mirror_config (`Dict[int, List[int]]`): Mapping dictionary: {source: [target1, target2...]}
+            target_config (`Dict[int, TargetConfig]`): Targets config
+            chat_mapping (`Dict[int, List[int]]`): Mapping dictionary: {source: [target1, target2...]}
             database (`Database`): Message IDs storage
             logger (`str` | `logging.Logger`, optional): Logger. Defaults to None.
         """
-        self._mirror_config = mirror_config
-        self._source_chats = list(mirror_config.keys())
+        self._target_config = target_config
+        self._chat_mapping = chat_mapping
+        self._source_chats = list(chat_mapping.keys())
         self._database = database
 
         if isinstance(logger, str):
@@ -256,23 +285,24 @@ class Mirroring(EventHandlers):
 
         self.add_event_handler(self.on_new_message,
                                events.NewMessage(chats=self._source_chats))
-        self.add_event_handler(self.on_album, events.Album(chats=self._source_chats))
+        self.add_event_handler(
+            self.on_album, events.Album(chats=self._source_chats))
 
         self.add_event_handler(self.on_edit_message,
-                                   events.MessageEdited(chats=self._source_chats))
+                               events.MessageEdited(chats=self._source_chats))
 
         self.add_event_handler(self.on_deleted_message,
-                                   events.MessageDeleted(chats=self._source_chats))
+                               events.MessageDeleted(chats=self._source_chats))
 
     def printable_config(self: 'MirrorTelegramClient') -> str:
         """Get printable mirror config"""
 
         mirror_mapping = '\n'.join(
-            [f'{f} -> {", ".join(map(str, to.to))}' for (f, to) in self._mirror_config.items()])
+            [f'{f} -> {", ".join(map(str, to))}' for (f, to) in self._chat_mapping.items()])
 
         return (
             f'Mirror mapping: \n{ mirror_mapping }\n'
-            f'Mirror config: { self._mirror_config }\n'
+            f'Mirror targets config: { self._target_config }\n'
             f'Using database: { self._database }\n'
         )
